Refactored all forms to use field level validation - Grouped all validators and async validators together in their own respective files - Refactored form components/containers to use field level validation - Removed redundant files
