+++ src/locales/en/translations.json
@@ -85,7 +85,6 @@
     "Page not found.": "",
     "No match for": "",
     "Go back home": "",
-    "This field is required": "",
     "User \"{{displayName}}\" replicated successfuly": "",
     "There was a problem replicating the user": "",
     "Username": "",
@@ -133,7 +132,6 @@
     "Group members": "",
     "Available user groups": "",
     "Managed user groups": "",
-    "Exceeds maximum character limit of 50": "",
     "You left group": "",
     "You joined group": "",
     "There was a problem updating your group membership": "",
@@ -169,20 +167,6 @@
     "Selected user groups": "",
     "Available dimension restrictions for data analytics": "",
     "Selected dimension restrictions for data analytics": "",
-    "A username should be at least 2 characters long": "",
-    "Username may not exceed 140 characters": "",
-    "Value should be a number": "",
-    "Value should be an integer": "",
-    "Value should be a positive integer": "",
-    "Value should be a nagative integer": "",
-    "Please provide a valid international phone number (+0123456789)": "",
-    "A user should have at least one User Role": "",
-    "Passwords do not match": "",
-    "Please provide a valid email address": "",
-    "Attribute value needs to be unique, value already taken.": "",
-    "There was a problem checking if this attribute value is unique": "",
-    "Username already taken": "",
-    "There was a problem whilst checking the availability of this username": "",
     "Replicate user": "",
     "Are you sure you want to disable two factor authentication for": "",
     "Succesfully disabled two factor authentication for": "",
@@ -203,17 +187,33 @@
     "User Management": "",
     "{{count}} selected": "",
     "{{count}} selected_plural": "",
-    "{{fieldDisplayName}} is already taken": "",
-    "Could not verify if this {{fieldDisplayName}} is unique": "",
+    "Something went wrong when processing your request.": "",
+    "Validating...": "",
+    "Are you sure you want to remove the {{type}} \"{{displayName}}\"?": "",
+    "{{type}} \"{{displayName}}\" removed successfully": "",
+    "There was a problem removing {{type}} \"{{displayName}}\"?": "",
+    "Exceeds maximum character limit of 50": "",
+    "Please provide a valid email address": "",
+    "Value should be an integer": "",
+    "Value should be a nagative integer": "",
+    "Value should be a number": "",
+    "Value should be a positive integer": "",
+    "Passwords do not match": "",
+    "This field is required": "",
+    "A username should be at least 2 characters long": "",
+    "Username may not exceed 140 characters": "",
+    "A user should have at least one User Role": "",
+    "Please provide a valid international phone number (+0123456789)": "",
     "Password should be at least 8 characters long": "",
     "Password should be no longer than 34 characters": "",
     "Password should contain at least one lowercase letter": "",
     "Password should contain at least one UPPERCASE letter": "",
     "Password should contain at least one number": "",
     "Password should have at least one special character": "",
-    "Something went wrong when processing your request.": "",
-    "Validating...": "",
-    "Are you sure you want to remove the {{type}} \"{{displayName}}\"?": "",
-    "{{type}} \"{{displayName}}\" removed successfully": "",
-    "There was a problem removing {{type}} \"{{displayName}}\"?": ""
+    "Attribute value needs to be unique, value already taken.": "",
+    "There was a problem checking if this attribute value is unique": "",
+    "Username already taken": "",
+    "There was a problem whilst checking the availability of this username": "",
+    "{{fieldDisplayName}} is already taken": "",
+    "Could not verify if this {{fieldDisplayName}} is unique": ""
 }
